+++ lib/kitabu/extensions/rouge.rb
@@ -2,7 +2,8 @@ module Rouge
   module Plugins
     module Redcarpet
       def rouge_formatter(lexer)
-        Formatters::HTML.new({css_class: "highlight #{lexer.tag}"}.merge(lexer.options))
+        options = lexer.respond_to?(:options) ? lexer.options : {}
+        Formatters::HTML.new({css_class: "highlight #{lexer.tag}"}.merge(options))
       end
     end
   end